requirements.txt
streamlit
requests
beautifulsoup4
pandas
PyGithub